Output interactive graphics in valid AMP HTML
This applies to interactive graphics that are displayed using iframes.
If the graphic source is insecure then we have to simply remove it. For
graphics which are served over HTTPS, we create an amp-iframe.

There's currently no way to test HTTPS iframes automatically, IG was
unable to find an article which used them. Regardless, this is an
improvement on the current behaviour for any existing articles which
have iframe-embedded interactive graphics.