Hide duplicated links in more info section
Since the image links to the same place as the first link below it, it is redundant and we can avoid having to give the image an alternative text.

aria-hidden="true" prevents it from appearing in the accessibility tree and treats this image as presenational.

tab-index=-1 is then needed to ensure that you cannot keyboard navigate to the link, without this you could focus the link and nothing would be announced in a screen reader. 🐿 v2.13.0